add some controler tests, refactor. Затрачено 0.2д

+++ src/main/java/ru/yandex/practicum/market/dto/CartDto.java
@@ -3,5 +3,5 @@ package ru.yandex.practicum.market.dto;
 import java.math.BigDecimal;
 import java.util.List;
 
-public record CartDto(List<ItemDto> items, BigDecimal total) {
+public record CartDto(List<ItemDto> items, Long total) {
 }
