+++ src/main/java/ru/yandex/practicum/market/services/OrderService.java
@@ -12,7 +12,7 @@ public class OrderService {
 
     private final ItemDto item1 = new ItemDto(1L, "Title 1", "Description 1", "", 100L, 1);
     private final List<ItemDto> items = List.of(item1);
-    private final OrderDto order = new OrderDto(1L, items, BigDecimal.valueOf(100L));
+    private final OrderDto order = new OrderDto(1L, items, 100L);
 
     public List<OrderDto> getOrders() {
         // TODO : use repository
